+++ backend/src/extract-profile/transform/helpers/getDurationInMonths.test.ts
