+++ backend/src/extract-profile/transform/helpers/getDurationInMonths.ts
