+++ backend/src/extract-profile/transform/helpers/getImage.test.ts
