+++ backend/src/extract-profile/transform/helpers/getImage.ts
