+++ backend/src/extract-profile/transform/helpers/getText.test.ts
