+++ backend/src/extract-profile/transform/helpers/getText.ts
