+++ backend/src/routes/getAssignments.ts
