+++ backend/src/routes/getCandidates.ts
