+++ backend/src/types/ResultsAPI.ts
