+++ database/Dockerfile
@@ -1,0 +1,2 @@
+FROM postgres:alpine
+COPY initialize.sql /docker-entrypoint-initdb.d/
